+++ dev/js-WebAssembly/Network.js
@@ -320,13 +320,13 @@ class Network {
             }
 
             if (this.state != "initialised") {
-                this.initLayers(data[0].input.length, (data[0].expected || data[0].output).length)
+                this.initLayers(data[0].input.length, data[0].expected.length)
             }
 
             const startTime = Date.now()
 
             const dimension = data[0].input.length
-            const itemSize = dimension + (data[0].expected || data[0].output).length
+            const itemSize = dimension + data[0].expected.length
             const itemsCount = itemSize * data.length
 
             if (log) {
@@ -338,8 +338,8 @@ class Network {
 
             for (let di=0; di<data.length; di++) {
 
-                if (!data[di].hasOwnProperty("input") || (!data[di].hasOwnProperty("expected") && !data[di].hasOwnProperty("output"))) {
-                    return void reject("Data set must be a list of objects with keys: 'input' and 'expected' (or 'output')")
+                if (!data[di].hasOwnProperty("input") || !data[di].hasOwnProperty("expected")) {
+                    return void reject("Data set must be a list of objects with keys: 'input' and 'expected'")
                 }
 
                 let index = itemSize*di
@@ -349,8 +349,8 @@ class Network {
                     index++
                 }
 
-                for (let ei=0; ei<(data[di].expected || data[di].output).length; ei++) {
-                    typedArray[index] = (data[di].expected || data[di].output)[ei]
+                for (let ei=0; ei<data[di].expected.length; ei++) {
+                    typedArray[index] = data[di].expected[ei]
                     index++
                 }
             }
@@ -387,8 +387,8 @@ class Network {
                             index++
                         }
 
-                        for (let ei=0; ei<(validation.data[di].expected || validation.data[di].output).length; ei++) {
-                            typedArray[index] = (validation.data[di].expected || validation.data[di].output)[ei]
+                        for (let ei=0; ei<validation.data[di].expected.length; ei++) {
+                            typedArray[index] = validation.data[di].expected[ei]
                             index++
                         }
                     }
@@ -512,7 +512,7 @@ class Network {
 
             const startTime = Date.now()
             const dimension = data[0].input.length
-            const itemSize = dimension + (data[0].expected || data[0].output).length
+            const itemSize = dimension + data[0].expected.length
             const itemsCount = itemSize * data.length
             const typedArray = new Float32Array(itemsCount)
 
@@ -525,8 +525,8 @@ class Network {
                     index++
                 }
 
-                for (let ei=0; ei<(data[di].expected || data[di].output).length; ei++) {
-                    typedArray[index] = (data[di].expected || data[di].output)[ei]
+                for (let ei=0; ei<data[di].expected.length; ei++) {
+                    typedArray[index] = data[di].expected[ei]
                     index++
                 }
             }
@@ -605,7 +605,7 @@ class Network {
     }
 
     static get version () {
-        return "3.1.0"
+        return "3.2.0"
     }
 }